model loading, better precision time delta, fps counter

+++ src/main/scala/dev/miikat/scalacraft/engine/Game.scala
@@ -3,6 +3,6 @@ package dev.miikat.scalacraft.engine
 import org.joml.*
 
 trait Game:
-  def updateState(glfwWindow: Long, camera: Camera, delta: Double, mouseDelta: Vector2f): Unit
+  def updateState(glfwWindow: Long, camera: Camera, delta: Long, mouseDelta: Vector2f): Unit
   def scene: Scene
   def init(): Unit
